pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
    "gunicorn>=25.1.0",
    "python-dotenv>=1.2.1",
    "rollbar>=1.3.0",
    "whitenoise>=6.11.0",
    "ruff>=0.9.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
]
[tool.black]
line-length = 79

[tool.ruff]
line-length = 79